feat: impl cyclefold::pp blank
